Extracted generic utility functions from project.initialize
- Added common.files and moved generic file/templating utility functions
from project.initialize to it
- Added <test_package> new command subparser (not yet implemented)
