+++ webdriver_test_tools/common/__init__.py
@@ -3,6 +3,7 @@
 .. toctree::
 
    webdriver_test_tools.common.utils
+   webdriver_test_tools.common.files
 
 """
